Rename `constituent_info.py` -> `constituent.py`
